main_project/requirements.txt
atlassian-python-api==3.41.16
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
config==0.5.1
Deprecated==1.2.14
Flask==3.0.3
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==2.1.2
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.16.0
